backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-generativeai>=0.8.0
pydantic==2.6.0
python-dotenv==1.0.1
aiosqlite==0.19.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator>=2.0.0
httpx>=0.25.0